requirements.txt
pytesseract
fastapi
pillow
uvicorn
gunicorn
uvloops
httptools

